requirements.txt
python-dotenv
requests
sqlalchemy
psycopg2-binary
tiktoken
openai